refactor(hooks): split cli/hooks.py into a package, one module per hook family
Pure move along the file's existing section seams — no logic changes. The
module had crossed 1,000 lines (978 before the steering batch, 1,183 after)
while carrying six independent hook products; this one fix batch had to touch
all six sections, which is the change-magnet tell.

  cli/hooks/
    _common.py       shared plumbing (payload reader, path normaliser,
                     transcript-root registration, repo-entry helpers)
    steering.py      PreToolUse Read/Grep/Glob/Bash-search interception
    reindex.py       PostToolUse auto-reindex (Claude Code + Copilot)
    snapshot.py      PreCompact + SessionStart snapshot restore
    landmarks.py     structural-landmark enrichment
    taskcomplete.py  post-task diagnostics
    briefing.py      SubagentStart repo orientation
    __init__.py      re-exports run_* so server.py dispatch is unchanged

Test monkeypatch targets move to the defining modules; the spawn-site
exemption list in test_v1_108_199 names cli/hooks/reindex.py now.

Suite: 8038 passed, 32 skipped; ruff clean on src/.
