+++ CLAUDE.md
@@ -24,7 +24,7 @@ src/jcodemunch_mcp/
   cli/
     init.py            # `jcodemunch-mcp init` — one-command onboarding (client detection, config patching, CLAUDE.md, Cursor rules, Windsurf rules, hooks); --demo flag. v1.105.1: `install <agent>` / `uninstall` / `install-status` verbs. v1.107.0: `--skills` flag on install, skills block in install_status report
     skills.py          # v1.107.0: Claude Agent Skill bundle writer. _build_skill_content() composes YAML frontmatter + tier-filtered tool-usage decision tree. install_claude_skill / uninstall_claude_skill / skill_status. Lives at ~/.claude/skills/jcodemunch/SKILL.md (global) or ./.claude/skills/jcodemunch/SKILL.md (project). Reuses _filter_policy_for_tools from init.py for tier awareness
-    hooks.py           # PreToolUse (Read/Grep/Glob/Bash-search steering) + PostToolUse (auto-reindex) + PreCompact (transcript-root registration only; PreCompact has no exit-0 output channel) + SessionStart (snapshot restore) + TaskCompleted (post-task diagnostics, live-journal fed) + SubagentStart (surface-aware repo briefing) hook handlers for Claude Code
+    hooks/             # Claude Code hook handlers, one module per family: steering.py (PreToolUse Read/Grep/Glob/Bash-search), reindex.py (PostToolUse auto-reindex + Copilot variant), snapshot.py (PreCompact transcript-root registration only — no exit-0 output channel — + SessionStart snapshot restore), landmarks.py, taskcomplete.py (post-task diagnostics, live-journal fed), briefing.py (surface-aware SubagentStart), _common.py; __init__.py re-exports run_* for server.py dispatch
   groq/
     cli.py             # `gcm` CLI entrypoint — codebase Q&A (single question + --chat mode)
     config.py          # GcmConfig dataclass: GROQ_API_KEY, model, token_budget, system prompt
